product

+++ database/migrations/2023_02_08_145035_create_product_skus_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_id');
             $table->string('name')->nullable();
             // $table->string('barcode')->unique()->nullable();
-            $table->unsignedInteger('price')->nullable();
+            $table->decimal('price',12)->nullable();
             $table->unsignedInteger('stock')->nullable();
             $table->timestamps();
         });
